requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
mysqlclient==2.1.0
Pillow==8.4.0
pytz==2021.3
reportlab==3.6.2
sqlparse==0.4.2